+++ pytsite/widget/_base.py
@@ -48,6 +48,7 @@ class Abstract(_ABC):
         self._child_sep = kwargs.get('child_sep', '')
         self._children = []  # type: _List[Abstract]
         self._children_uids = []  # type: _List[str]
+        self._group_wrap = True
 
         # Check validation rules
         if not isinstance(self._rules, (list, tuple)):
@@ -68,13 +69,13 @@ class Abstract(_ABC):
             self._value = _deepcopy(self._default)
 
     @_abstractmethod
-    def get_html_em(self, **kwargs) -> _html.Element:
-        """Get an HTML element representation of the widget.
+    def _get_element(self, **kwargs) -> _html.Element:
+        """Hook.
         """
         pass
 
-    def render(self, **kwargs) -> str:
-        """Render the widget into a string.
+    def get_element(self, **kwargs) -> _html.Element:
+        """Get an HTML element representation of the widget.
         """
         # Wrapper div
         self._wrap_em.set_attr('data_cid', self.__module__ + '.' + self.__class__.__name__)
@@ -101,13 +102,13 @@ class Abstract(_ABC):
             self._wrap_em.set_attr('data_replaces', self._replaces)
 
         # Get widget's HTML element
-        html_em = self.get_html_em(**kwargs)
+        em = self._get_element(**kwargs)
 
         # Wrapper CSS
         wrap_css = 'pytsite-widget widget-uid-{} {}'.format(self._uid, self._css)
         if self._hidden:
             wrap_css += ' hidden'
-        if isinstance(html_em, _html.TagLessElement) and not html_em.content:
+        if isinstance(em, _html.TagLessElement) and not em.content:
             wrap_css += ' empty'
         self._wrap_em.set_attr('cls', wrap_css)
 
@@ -116,10 +117,19 @@ class Abstract(_ABC):
             for k, v in self._data.items():
                 self._wrap_em.set_attr('data_' + k, v)
 
-        # Wrap widget's HTML
-        self._wrap_em.append(html_em)
+        # Wrap into 'form-group' div
+        if self._group_wrap:
+            em = self._wrap_into_group(em)
 
-        return self._wrap_em.render()
+        # Wrap widget's HTML
+        self._wrap_em.append(em)
+
+        return self._wrap_em
+
+    def render(self, **kwargs) -> str:
+        """Render the widget into a string.
+        """
+        return self.get_element(**kwargs).render()
 
     def __str__(self) -> str:
         return self.render()
@@ -377,6 +387,14 @@ class Abstract(_ABC):
         """
         return self._children
 
+    @property
+    def group_wrap(self) -> bool:
+        return self._group_wrap
+
+    @group_wrap.setter
+    def group_wrap(self, value: bool) -> bool:
+        self._group_wrap = value
+
     def has_child(self, uid: str) -> bool:
         """Check if the widget has a child.
         """
@@ -466,7 +484,7 @@ class Abstract(_ABC):
         for rule in self.get_rules():
             rule.validate(self.get_val(mode='validation'))
 
-    def _group_wrap(self, content) -> _html.Element:
+    def _wrap_into_group(self, content) -> _html.Element:
         """Wraps a widget, an HTML element or a string into 'form-group' container.
 
         :type content: pytsite.widget._base.Abstract | pytsite.html.Element | str
@@ -474,7 +492,7 @@ class Abstract(_ABC):
         if isinstance(content, str):
             content = _html.TagLessElement(content)
         elif isinstance(content, Abstract):
-            content = content.get_html_em()
+            content = content._get_element()
 
         if self._h_size:
             content = content.wrap(_html.Div(cls='h-sizer ' + self._h_size))
@@ -522,10 +540,184 @@ class Container(Abstract):
 
         self._css += ' widget-container'
 
-    def get_html_em(self, **kwargs) -> _html.Element:
+    def validate(self):
+        """Validate widget's rules.
+        """
+        for w in self.children:
+            for rule in w.get_rules():
+                rule.validate(w.get_val(mode='validation'))
+
+    def _get_element(self, **kwargs) -> _html.Element:
         cont = _html.TagLessElement(child_sep=self._child_sep)
 
         for w in self.children:
-            cont.append(_html.TagLessElement(w.render()))
+            cont.append(w.get_element())
 
         return cont
+
+
+class MultiRow(Abstract):
+    """Multi Row Widget.
+    """
+
+    def __init__(self, uid: str, **kwargs):
+        self._add_btn_title = kwargs.get('add_btn_title', _lang.t('pytsite.widget@append'))
+
+        super().__init__(uid, **kwargs)
+
+        self.css += ' widget-multi-row'
+        self.assets.append('pytsite.widget@css/multi-row.css')
+        self.assets.append('pytsite.widget@js/multi-row.js')
+
+    def append_child(self, widget: Abstract):
+        raise NotImplementedError('This widget can not contain children')
+
+    @property
+    def children(self):
+        raise NotImplementedError('This widget can not contain children')
+
+    def set_val(self, value: list, **kwargs):
+        if value is None:
+            value = []
+
+        # If value comes from HTTP input, it usually would be a dict, and it should be converted to a list
+        if isinstance(value, dict):
+            new_val = []
+            keys = list(value.keys())
+            val_len = len(value[keys[0]])
+            for n in range(0, val_len):
+                new_val_item = {}
+                for k in keys:
+                    new_val_item[k] = value[k][n]
+                new_val.append(new_val_item)
+
+            value = new_val
+
+        # Check type of the entire value
+        if not isinstance(value, list):
+            raise TypeError('List expected, {} given'.format(type(value)))
+
+        # Cleanup value
+        clean_value = []
+        for v in value:
+            if not isinstance(v, dict):
+                raise TypeError('Dict expected, {} given'.format(type(v)))
+
+            # Check that all values of the dict is not empty
+            clean_v_values = [v_value for v_value in v.values() if v_value]
+            if clean_v_values:
+                clean_value.append(v)
+        value = clean_value
+
+        # Create child widgets based on value
+        self._children = []
+        for value_item in value:
+            children_row = []
+            for w in self._get_widgets_row():
+                if not isinstance(w, Abstract):
+                    raise TypeError('Widget expected, {} given'.format(type(w)))
+
+                w.value = value_item[w.name] if w.name in value_item else None
+                children_row.append(w)
+
+            self._children.append(children_row)
+
+        super().set_val(value)
+
+    def validate(self):
+        """Validate widget's rules.
+        """
+        row_i = 0
+        for row in self._children:
+            widget_i = 0
+            for w in row:
+                for rule in w.get_rules():
+                    try:
+                        rule.validate(w.get_val(mode='validation'))
+                    except _validation.error.RuleError as e:
+                        msg_id = 'pytsite.widget@multi_row_validation_error'
+                        msg_args = {
+                            'row_index': row_i + 1,
+                            'widget_title': self._get_headers_row()[widget_i],
+                            'orig_msg': str(e)
+                        }
+
+                        raise _validation.error.RuleError(msg_id, msg_args)
+
+                widget_i += 1
+
+            row_i += 1
+
+    @_abstractmethod
+    def _get_headers_row(self) -> list:
+        """Hook.
+        """
+        pass
+
+    @_abstractmethod
+    def _get_widgets_row(self) -> list:
+        """Hook.
+        """
+        pass
+
+    def _get_element(self, **kwargs) -> _html.Element:
+        def _build_row(widgets: _List[Abstract], i: int = 0, add_cls: str='') -> _html.Tr:
+            slot_tr = _html.Tr(cls='slot ' + add_cls)
+            slot_tr.append(_html.Td('[{}]'.format(i + 1), cls='order-col'))
+
+            # Widgets
+            for w in widgets:
+                w.name = '{}[{}][]'.format(self.name, w.name)
+                w.group_wrap = False
+                w.css += ' widget-row-col'
+
+                w_td = _html.Td(cls='widget-col')
+                w_td.append(w.get_element())
+
+                slot_tr.append(w_td)
+
+            # Actions
+            actions_td = _html.Td(cls='actions-col')
+            remove_btn = _html.A(href='#', cls='button-remove-slot btn btn-xs btn-danger')
+            remove_btn.append(_html.I(cls='fa fa-icon fa-remove'))
+            actions_td.append(remove_btn)
+            slot_tr.append(actions_td)
+
+            return slot_tr
+
+        table = _html.Table(cls='content-table')
+
+        # Header
+        thead = _html.THead(cls='hidden slots-header')
+        table.append(thead)
+        row = _html.Tr()
+        thead.append(row)
+        row.append(_html.Th('#', cls='order-col'))
+        for v in self._get_headers_row():
+            row.append(_html.Th(v, cls='widget-col'))
+        row.append(_html.Th(cls='widget-col'))
+
+        # Table body
+        tbody = _html.TBody(cls='slots')
+        table.append(tbody)
+
+        # Sample slot
+        sample_row = self._get_widgets_row()
+        tbody.append(_build_row(sample_row, add_cls='sample hidden'))
+
+        # Rows
+        for i in range(0, len(self._children)):
+            tbody.append(_build_row(self._children[i], i))
+
+        # Footer
+        tfoot = _html.TFoot()
+        tr = _html.Tr()
+        td = _html.Td(colspan=len(self._get_widgets_row()) + 2)
+        add_btn = _html.A(self._add_btn_title or '', href='#', cls='button-add-slot btn btn-default btn-xs')
+        add_btn.append(_html.I(cls='fa fa-plus'))
+        td.append(add_btn)
+        tr.append(td)
+        tfoot.append(tr)
+        table.append(tfoot)
+
+        return table